feat: Add GPIO class for controlling LEDs
This commit adds a new GPIO class to the project, which provides methods for controlling the state of LEDs connected to GPIO pins. The GPIO class is implemented in the GPIO.h file and includes necessary header files and dependencies.

The GPIO class constructor takes the path to the device file as a parameter and initializes the file descriptor for the GPIO device. It also initializes the GPIO handle request and GPIO handle data structures.

The GPIO class includes methods for setting the state of LED1 and LED2. These methods take an integer parameter representing the desired state of the LED (0 for off, 1 for on) and update the LED state accordingly.

This commit is part of the ongoing development of the UART/GPIO user space driver.